msRythmBot/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
autopep8==1.5.7
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
discord==1.0.1
discord.py==1.7.1
idna==2.10
multidict==5.1.0
pycodestyle==2.7.0
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.17.1
requests==2.25.1
six==1.16.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.5
yarl==1.6.3
youtube-dl==2021.6.6
